Only use this for profiles with MFA
We want to use duration_seconds for longer sessions but AWS limits sessions to 1 hour when chaining sessions as done here. The error was:

The requested DurationSeconds exceeds the 1 hour session limit for roles assumed by role chaining.

So we're going to stop using this hack for roles without MFA (it's only useful for MFA profiles anyway) and then longer sessions should work as normal.